Moved github latest tag fetching to separate syncronous file
